refactoring to debug out to AT_DIAGNOSTIC
